post app for testing

+++ post_blog/views.py
@@ -2,6 +2,7 @@ from django.shortcuts import render, redirect, get_object_or_404
 from .models import post_model
 from .forms import post_model_form
 
+
 def home_page(request):
     post_models = post_model.objects.all()
 
@@ -19,10 +20,15 @@ def post_details(request, slug, id):
 
 
 def create_post_view(request):
+    if not request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
         form = post_model_form(request.POST)
         if form.is_valid():
-            form.save()
+            post = form.save(commit=False)
+            post.author = request.user
+            post.save()
             return redirect('home')
     else:
         form = post_model_form()
@@ -32,6 +38,9 @@ def create_post_view(request):
 
 
 def edit_post(request, post_slug):
+    if not request.user.is_authenticated:
+        return redirect('home')
+
     # post = get_object_or_404(post_model, id=post_slug)
     post = post_model.objects.get(slug=post_slug)
 
@@ -51,6 +60,9 @@ def edit_post(request, post_slug):
 
 
 def post_delete(request, post_id):
+    if not request.user.is_authenticated:
+        return redirect('home')
+
     post = get_object_or_404(post_model, id=post_id)
     if request.method == 'POST':
         post.delete()
